remove unused, refactor
